Split the page into separate files
